+++ modules/components/src/main/java/io/github/palexdev/materialfx/controls/buttons/MFXButton.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Parisi Alessandro - alessandro.parisi406@gmail.com
+ * Copyright (C) 2023 Parisi Alessandro - alessandro.parisi406@gmail.com
  * This file is part of MaterialFX (https://github.com/palexdev/MaterialFX)
  *
  * MaterialFX is free software: you can redistribute it and/or
@@ -16,7 +16,7 @@
  * along with MaterialFX. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.palexdev.materialfx.controls;
+package io.github.palexdev.materialfx.controls.buttons;
 
 import io.github.palexdev.materialfx.behaviors.MFXButtonBehavior;
 import io.github.palexdev.materialfx.controls.base.MFXLabeled;
@@ -30,6 +30,22 @@ import javafx.scene.Node;
 import java.util.List;
 import java.util.function.Supplier;
 
+/**
+ * Custom implementation of a button which extends {@link MFXLabeled}, has its own skin
+ * {@link MFXButtonSkin} and its own behavior {@link MFXButtonBehavior}.
+ * <p></p>
+ * {@code MFXButton} has 5 variants that mainly override the {@link #defaultStyleClasses()} method
+ * to allow themes to style them according to their style classes:
+ * <p> - {@link MFXElevatedButton}
+ * <p> - {@link MFXFilledButton}
+ * <p> - {@link MFXTonalFilledButton}
+ * <p> - {@link MFXOutlinedButton}
+ * <p> - {@link MFXTextButton}
+ * <p></p>
+ * <b>This base class is un-styled by the default official themes, a perfect start to implement custom styled buttons.</b>
+ * <p>
+ * The default style class of this component is: '.mfx-button'.
+ */
 public class MFXButton extends MFXLabeled<MFXButtonBehavior> {
 	//================================================================================
 	// Properties
@@ -59,6 +75,45 @@ public class MFXButton extends MFXLabeled<MFXButtonBehavior> {
 	}
 
 	//================================================================================
+	// Variants
+	//================================================================================
+
+	/**
+	 * @return a new {@link MFXElevatedButton}
+	 */
+	public static MFXElevatedButton elevated() {
+		return new MFXElevatedButton();
+	}
+
+	/**
+	 * @return a new {@link MFXFilledButton}
+	 */
+	public static MFXFilledButton filled() {
+		return new MFXFilledButton();
+	}
+
+	/**
+	 * @return a new {@link MFXTonalFilledButton}
+	 */
+	public static MFXTonalFilledButton tonalFilled() {
+		return new MFXTonalFilledButton();
+	}
+
+	/**
+	 * @return a new {@link MFXOutlinedButton}
+	 */
+	public static MFXOutlinedButton outlined() {
+		return new MFXOutlinedButton();
+	}
+
+	/**
+	 * @return a new {@link MFXTextButton}
+	 */
+	public static MFXTextButton text() {
+		return new MFXTextButton();
+	}
+
+	//================================================================================
 	// Methods
 	//================================================================================
 	private void initialize() {
@@ -66,6 +121,10 @@ public class MFXButton extends MFXLabeled<MFXButtonBehavior> {
 		setDefaultBehaviorProvider();
 	}
 
+	/**
+	 * If not disabled, fires a new {@link ActionEvent}, triggering the {@link EventHandler} specified
+	 * by the {@link #onActionProperty()}.
+	 */
 	public void fire() {
 		if (!isDisabled()) fireEvent(new ActionEvent());
 	}
@@ -89,33 +148,15 @@ public class MFXButton extends MFXLabeled<MFXButtonBehavior> {
 	}
 
 	//================================================================================
-	// Variants
-	//================================================================================
-	public static class MFXElevatedButton extends MFXButton {
-		public MFXElevatedButton() {
-		}
-
-		public MFXElevatedButton(String text) {
-			super(text);
-		}
-
-		public MFXElevatedButton(String text, Node icon) {
-			super(text, icon);
-		}
-
-		@Override
-		public List<String> defaultStyleClasses() {
-			return List.of("mfx-button", "elevated");
-		}
-	}
-
-	//================================================================================
 	// Getters/Setters
 	//================================================================================
 	public EventHandler<ActionEvent> getOnAction() {
 		return onAction.get();
 	}
 
+	/**
+	 * Specifies the action to execute when an {@link ActionEvent} is fired on this button.
+	 */
 	public EventHandlerProperty<ActionEvent> onActionProperty() {
 		return onAction;
 	}