+++ modules/components/src/main/java/module-info.java
@@ -2,9 +2,19 @@ module mfx.components {
 	requires transitive javafx.controls;
 
 	requires transitive mfx.core;
+	requires transitive mfx.effects;
 	requires transitive mfx.localization;
 	requires transitive mfx.resources;
 
+	// Behaviors
 	exports io.github.palexdev.materialfx.behaviors;
-	exports io.github.palexdev.materialfx.controls;
+
+	// Controls
+	exports io.github.palexdev.materialfx.controls.buttons;
+
+	// Skins
+	exports io.github.palexdev.materialfx.skins;
+
+	// Theming
+	exports io.github.palexdev.materialfx.theming;
 }